Split FST logic into another file in dvd lib
