+++ lib/dolphin/dvd/dvd.cpp
@@ -15,45 +15,32 @@
 #include <utility>
 #include <vector>
 
+#include "dvd.hpp"
+
 #include "../../internal.hpp"
 
+using namespace aurora::dvd::impl;
+
+namespace aurora::dvd::impl {
+  NodHandle* s_partition = nullptr;
+  std::vector<FSTEntry> s_fstEntries;
+  // Map from "virtual" FST entryNums (matching base disc, game-assigned for new overlay files)
+  // To the "physical" FST entryNums (that we use for navigating the tree).
+  // Unfilled spots are given the k_invalidFstEntry value.
+  std::vector<PhysicalEntryNum> s_fstEntryMap;
+  s32 s_baseEntryCount;
+  PhysicalEntryNum s_currentDir = 0;
+  std::string s_currentPath = "/";
+  BOOL s_autoInvalidation = FALSE;
+  BOOL s_autoFatalMessaging = FALSE;
+  DVDDiskID s_diskID = {};
+  DVDLowCallback s_resetCoverCallback = nullptr;
+  bool s_initialized = false;
+  bool s_overlayCallbacksSet = false;
+  AuroraOverlayCallbacks s_overlayCallbacks;
+}
+
 namespace {
-
-aurora::Module Log("aurora::dvd");
-
-using PhysicalEntryNum = s32;
-using VirtualEntryNum = s32;
-constexpr s32 k_invalidFstEntry = -1;
-
-struct FSTEntry {
-  std::string name;
-  bool isDir = false;
-  PhysicalEntryNum parent = 0;
-  u32 nextOrLength = 0;
-  void* overlayData = nullptr;
-  // Original entry num on the base game disc, BEFORE being re-organized by overlays.
-  VirtualEntryNum origEntryNum = 0;
-};
-
-struct IterateNode {
-  std::string name;
-  bool isDir;
-  u32 originalEntryNum;
-  u32 size;
-  void* overlayData;
-  std::vector<std::shared_ptr<IterateNode>> children;
-
-  IterateNode(std::string name, bool isDir, u32 size, u32 originalEntryNum, void* overlayData)
-  : name(std::move(name)), isDir(isDir), size(size), originalEntryNum(originalEntryNum), overlayData(overlayData) {}
-
-  IterateNode(std::string name, bool isDir, u32 size, u32 originalEntryNum)
-  : name(std::move(name)), isDir(isDir), size(size), originalEntryNum(originalEntryNum), overlayData(nullptr) {}
-};
-
-struct IterateContext {
-  std::shared_ptr<IterateNode> root;
-  std::vector<std::pair<std::shared_ptr<IterateNode>, u32>> dirStack;
-};
 
 class CommandDataBase {
 public:
@@ -79,24 +66,6 @@ public:
   }
 };
 
-CommandDataNod* s_disc;
-NodHandle* s_partition = nullptr;
-std::vector<FSTEntry> s_fstEntries;
-// Map from "virtual" FST entryNums (matching base disc, game-assigned for new overlay files)
-// To the "physical" FST entryNums (that we use for navigating the tree).
-// Unfilled spots are given the k_invalidFstEntry value.
-std::vector<PhysicalEntryNum> s_fstEntryMap;
-s32 s_baseEntryCount;
-PhysicalEntryNum s_currentDir = 0;
-std::string s_currentPath = "/";
-BOOL s_autoInvalidation = FALSE;
-BOOL s_autoFatalMessaging = FALSE;
-DVDDiskID s_diskID = {};
-DVDLowCallback s_resetCoverCallback = nullptr;
-bool s_initialized = false;
-bool s_overlayCallbacksSet = false;
-AuroraOverlayCallbacks s_overlayCallbacks;
-
 class CommandDataOverlay final : public CommandDataBase {
 public:
   void* handle;
@@ -114,14 +83,7 @@ public:
   }
 };
 
-struct OverlayFileEntry {
-  std::string fileName;
-  void* userData;
-  u32 size;
-  s32 entryNum;
-};
-
-std::vector<OverlayFileEntry> s_overlayFiles;
+CommandDataNod* s_disc;
 
 void clearState() {
   if (s_partition != nullptr) {
@@ -192,202 +154,8 @@ void sdlStreamClose(void* userData) {
   SDL_CloseIO(io);
 }
 
-u32 fstCallback(u32 index, NodNodeKind kind, const char* name, u32 size, void* userData) {
-  auto* ctx = static_cast<IterateContext*>(userData);
-
-  while (index >= ctx->dirStack.back().second) {
-    ctx->dirStack.pop_back();
-  }
-
-  const auto newEntry = std::make_shared<IterateNode>(
-    name,
-    (kind == NOD_NODE_KIND_DIRECTORY),
-    size,
-    index);
-
-  const auto& curDir = ctx->dirStack.back().first;
-  curDir->children.push_back(newEntry);
-
-  if (newEntry->isDir) {
-    ctx->dirStack.emplace_back(newEntry, size);
-  }
-
-  return index + 1;
-}
-
-bool nameEqualsIgnoreCase(std::string_view lhs, std::string_view rhs);
-
-IterateNode* findNode(const IterateNode& node, const std::string_view name) {
-  for (const auto& child : node.children) {
-    if (nameEqualsIgnoreCase(child->name, name)) {
-      return child.get();
-    }
-  }
-
-  return nullptr;
-}
-
-void mergeOverlayFileIntoContext(const IterateContext& context, const OverlayFileEntry& overlayFile) {
-  if (overlayFile.entryNum < s_baseEntryCount) {
-    Log.error(
-      "Overlay file {} has entryNum {} which is already used by the base disc!",
-      overlayFile.fileName,
-      overlayFile.entryNum);
-    return;
-  }
-
-  IterateNode* node = context.root.get();
-  std::string_view filePath = overlayFile.fileName;
-
-  assert(filePath.starts_with('/'));
-  filePath = filePath.substr(1);
-  while (true) {
-    const auto nextDelim = filePath.find('/');
-    if (nextDelim == std::string_view::npos) {
-      break;
-    }
-
-    const auto segment = filePath.substr(0, nextDelim);
-    filePath = filePath.substr(nextDelim + 1);
-
-    const auto existingNode = findNode(*node, segment);
-    if (existingNode) {
-      if (!existingNode->isDir) {
-        Log.error("Overlay file {} needs directory that's already a file!", overlayFile.fileName);
-        return;
-      }
-
-      node = existingNode;
-    } else {
-      const auto newNode = std::make_shared<IterateNode>(std::string(segment), true, 0, k_invalidFstEntry);
-      node->children.push_back(newNode);
-      node = newNode.get();
-    }
-  }
-
-  // Remainder of fileName is the actual file name, and node is the directory we're in.
-
-  auto newNode = IterateNode(std::string(filePath), false, overlayFile.size, overlayFile.entryNum, overlayFile.userData);
-  const auto existingNode = findNode(*node, filePath);
-  if (existingNode) {
-    if (existingNode->isDir) {
-      Log.error("Overlay file {} overlaps directory with same name!", overlayFile.fileName);
-      return;
-    }
-
-    newNode.originalEntryNum = existingNode->originalEntryNum;
-
-    // Replace existing disc entry.
-    *existingNode = std::move(newNode);
-  } else {
-    // Add new entry.
-    Log.debug("Adding new entry num: {} -> {}", overlayFile.entryNum, overlayFile.fileName);
-    newNode.originalEntryNum = overlayFile.entryNum;
-
-    node->children.emplace_back(std::make_shared<IterateNode>(std::move(newNode)));
-  }
-}
-
-void mergeOverlayFilesIntoContext(const IterateContext& context) {
-  for (const auto& overlayFile : s_overlayFiles) {
-    mergeOverlayFileIntoContext(context, overlayFile);
-  }
-}
-
-void makeFstRecursive(IterateNode& node, u32 parent) {
-  if (node.originalEntryNum != k_invalidFstEntry) {
-    if (s_fstEntryMap.size() <= node.originalEntryNum) {
-      s_fstEntryMap.resize(node.originalEntryNum + 1, k_invalidFstEntry);
-    }
-
-    auto& map = s_fstEntryMap[node.originalEntryNum];
-    if (map != k_invalidFstEntry) {
-      Log.error("File {} with virtual entry num {} already exists in map!", node.name, node.originalEntryNum);
-      return;
-    }
-
-    map = static_cast<PhysicalEntryNum>(s_fstEntries.size());
-  }
-
-  if (!node.isDir) {
-    assert(node.children.empty());
-    assert(node.originalEntryNum != k_invalidFstEntry);
-
-    s_fstEntries.emplace_back(node.name, false, parent, node.size, node.overlayData, node.originalEntryNum);
-    return;
-  }
-
-  std::ranges::sort(node.children, [](const auto& a, const auto& b) { return a->name < b->name; });
-
-  const auto ourIndex = s_fstEntries.size();
-  s_fstEntries.emplace_back(node.name, true, parent, 0, node.overlayData, node.originalEntryNum);
-
-  for (const auto& child : node.children) {
-    makeFstRecursive(*child, ourIndex);
-  }
-
-  s_fstEntries[ourIndex].nextOrLength = s_fstEntries.size();
-}
-
-void makeFstFromContext(const IterateContext& context) {
-  makeFstRecursive(*context.root, 0);
-}
-
-s32 calcEntryCount(const IterateNode& node) {
-  s32 counter = 1;
-
-  for (const auto& child : node.children) {
-    counter += calcEntryCount(*child);
-  }
-
-  return counter;
-}
-
-bool rebuildFST() {
-  using namespace std::string_literals;
-
-  if (s_partition == nullptr) {
-    return false;
-  }
-
-  // TODO: Ensure current dir still valid after rebuild.
-
-  s_fstEntries.clear();
-  s_fstEntryMap.clear();
-  IterateContext ctx;
-  ctx.root = std::make_shared<IterateNode>(""s, true, 0, static_cast<u32>(0));
-  ctx.dirStack.emplace_back(ctx.root, std::numeric_limits<u32>::max());
-
-  nod_partition_iterate_fst(s_partition, fstCallback, &ctx);
-  s_baseEntryCount = calcEntryCount(*ctx.root);
-  mergeOverlayFilesIntoContext(ctx);
-  makeFstFromContext(ctx);
-
-  return true;
-}
-
-bool nameEqualsIgnoreCase(const std::string_view lhs, const std::string_view rhs) {
-  if (lhs.size() != rhs.size()) {
-    return false;
-  }
-  for (size_t i = 0; i < rhs.size(); ++i) {
-    char lc = lhs[i];
-    char rc = rhs[i];
-    if (lc >= 'a' && lc <= 'z') {
-      lc = static_cast<char>(lc - 'a' + 'A');
-    }
-    if (rc >= 'a' && rc <= 'z') {
-      rc = static_cast<char>(rc - 'a' + 'A');
-    }
-    if (lc != rc) {
-      return false;
-    }
-  }
-  return true;
-}
-
 bool nameEqualsIgnoreCase(const std::string& lhs, const char* rhs, size_t rhsLen) {
-  return nameEqualsIgnoreCase(lhs, std::string_view(rhs, rhsLen));
+  return aurora::dvd::impl::nameEqualsIgnoreCase(lhs, std::string_view(rhs, rhsLen));
 }
 
 PhysicalEntryNum findInDir(PhysicalEntryNum dirEntry, const char* name, size_t nameLen) {
@@ -615,56 +383,6 @@ bool aurora_dvd_open(const char* disc_path) {
 }
 
 void aurora_dvd_close(void) { clearState(); }
-
-static bool validateOverlayFile(const AuroraOverlayFile& file) {
-  const std::string_view name(file.fileName);
-
-  if (!name.starts_with('/')) {
-    Log.error("Overlay path {} does not start with /", name);
-    return false;
-  }
-
-  if (file.size > std::numeric_limits<u32>::max()) {
-    Log.error("Overlay file sizes above 4 GiB are not supported: {}", name);
-    return false;
-  }
-
-  if (file.entryNum < 0) {
-    Log.error("Overlay file entry is below zero: {}", name);
-    return false;
-  }
-
-  return true;
-}
-
-s32 aurora_dvd_base_entry_count() {
-  return s_baseEntryCount;
-}
-
-void aurora_dvd_overlay_files(const AuroraOverlayFile* files, size_t nFiles) {
-  if (!s_overlayCallbacksSet) {
-    Log.fatal("aurora_dvd_overlay_callbacks not called before aurora_dvd_overlay_files!");
-  }
-
-  s_overlayFiles.clear();
-
-  for (size_t i = 0; i < nFiles; i++) {
-    const auto& file = files[i];
-
-    if (!validateOverlayFile(file)) {
-      continue;
-    }
-
-    s_overlayFiles.emplace_back(file.fileName, file.userData, static_cast<u32>(file.size), file.entryNum);
-  }
-
-  rebuildFST();
-}
-
-void aurora_dvd_overlay_callbacks(const AuroraOverlayCallbacks* callbacks) {
-  s_overlayCallbacks = *callbacks;
-  s_overlayCallbacksSet = true;
-}
 
 void DVDInit(void) {}
